+++ VDA-StorageNode-Registry/scripts/deploy.ts
@@ -1,14 +1,15 @@
 import hre from "hardhat";
 import { deploy } from "./libraries/deployment";
-import { saveABI, saveDeployedAddress } from "./libraries/utils";
+import { saveABI, saveDeployArgument, saveDeployedAddress } from "./libraries/utils";
 
 async function main() {
-  const { diamondAddress, tokenAddress, facetsAddress, abi } = await deploy(
+  const { diamondAddress, tokenAddress, facetsAddress, abi, deployArgument } = await deploy(
     undefined, 
     ['VDAVerificationFacet', 'VDADataCenterFacet', 'VDAStorageNodeFacet', 'VDAStorageNodeManagementFacet']);
 
   await saveDeployedAddress(hre.network.name, diamondAddress, tokenAddress, facetsAddress);
   await saveABI(abi);
+  await saveDeployArgument(deployArgument);
 }  
  
 // We recommend this pattern to be able to use async/await everywhere